Minor refactor of existing Contexts
